Moved find_versions to Linter. refs: #101
