+++ moncic/lint.py
@@ -4,6 +4,7 @@ import contextlib
 import re
 from collections import defaultdict
 from functools import cached_property
+from pathlib import Path
 from typing import TYPE_CHECKING, Optional
 
 import git
@@ -29,6 +30,54 @@ class Linter(contextlib.ExitStack):
 
     def warning(self, message: str):
         print(message)
+
+    def find_versions(self) -> dict[str, str]:
+        """
+        Get the program version from sources.
+
+        Return a dict mapping version type to version strings
+        """
+        versions: dict[str, str] = {}
+
+        path = Path(self.source.source.path)
+        if (autotools := path / "configure.ac").exists():
+            re_autotools = re.compile(r"\s*AC_INIT\s*\(\s*[^,]+\s*,\s*\[?([^,\]]+)")
+            with autotools.open("rt") as fd:
+                for line in fd:
+                    if (mo := re_autotools.match(line)):
+                        versions["autotools"] = mo.group(1).strip()
+                        break
+
+        if (meson := path / "meson.build").exists():
+            re_meson = re.compile(r"\s*project\s*\(.+version\s*:\s*'([^']+)'")
+            with meson.open("rt") as fd:
+                for line in fd:
+                    if (mo := re_meson.match(line)):
+                        versions["meson"] = mo.group(1).strip()
+                        break
+
+        if (cmake := path / "CMakeLists.txt").exists():
+            re_cmake = re.compile(r"""\s*set\s*\(\s*PACKAGE_VERSION\s+["']([^"']+)""")
+            with cmake.open("rt") as fd:
+                for line in fd:
+                    if (mo := re_cmake.match(line)):
+                        versions["cmake"] = mo.group(1).strip()
+                        break
+
+        if (news := path / "NEWS.md").exists():
+            re_news = re.compile(r"# New in version (.+)")
+            with news.open("rt") as fd:
+                for line in fd:
+                    if (mo := re_news.match(line)):
+                        versions["news"] = mo.group(1).strip()
+                        break
+
+        # TODO: check setup.py
+        # TODO: can it be checked without checking out the branch and executing it?
+        # TODO: check debian/changelog in a subclass
+        # TODO: check specfile in a subclass
+
+        return versions
 
     def check_local_remote_sync(self, name: str) -> str:
         """
@@ -198,7 +247,7 @@ class Linter(contextlib.ExitStack):
         lint-check the sources, using analyzer to output results
         """
         # Check for version mismatches
-        versions = self.source.find_versions()
+        versions = self.find_versions()
 
         by_version: dict[str, list[str]] = defaultdict(list)
         for name, version in versions.items():
